+++ CHANGELOG.md
@@ -31,6 +31,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 * `Removed` VNet Edge. Connecting OpenVNet to other networks on the L2 level should now be done by using `promiscuous interface mode` instead.
 
+* `Removed` Security groups. Use filters instead for firewall functionality.
+
 ## [0.9] - 2016-04-19
 
 * `Added` A new `topology manager` that automatically creates `datapath_network` and `datapath_route_link` entries in the database.